revert config changes
